#108 - `UserEnrollmentType` move and usage
Move `UserEnrollmentType` from `api.section.users.dto.ts` to `canvas.interfaces.ts` and also use it as type of the enrollment `type` property.
